backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
python-magic-bin==0.4.14; sys_platform == "win32"
python-magic==0.4.27; sys_platform != "win32"
opencv-python-headless==4.10.0.84
Pillow==11.1.0
httpx==0.28.1
numpy>=2.1.0
paddleocr>=2.9.1
paddlepaddle==3.3.0
pypdfium2>=4.0.0
pytest==8.3.4
pytest-asyncio==0.25.0
